Disable check for selection
When the editor has selected text, I didn't allow the package to run.
This make things easier in the beginning because I couldn't get it to
work right when there was a selection.  However, this seems to be useless
now.

In addition it prevents users to backspace to a tag to activate the
package.  This is because to backspace, Atom selects to the left then
deletes.  Since there is a selection when backspacing, the package
doesn't activate.

Fixes #1