Add support for html tree sitter scope name
The html package is using a new name and method to generate the scopes.
One change is a new name for html tags: 'entity.tag.name'.  Another change
is the tag scope no longer activates when the cursor is on the right side of
then tag.  Not sure if bug.