Fix redo for start tag sync
When running undo, the tag changes aren't moved to the redo
stack. So redoing does nothing. This package is seeing the undo as
a tag change, so it tries to the sync it to the end tag.  However, the
undo operation already changed the tag back. This must mess up the undo
stack.  So I ignore undo by resetting if the tag isn't changed.

This works for end tag sync, but it causes a dormant bug to activate.
There is a race condition when editing end tags.  The tag changes are
doubling up.